+++ skythought/tools/tasks/__init__.py
@@ -1,0 +1,46 @@
+from .aime.aime_handler import AIMETaskHandler
+from .amc23.amc23_handler import AMC23TaskHandler
+from .apps.apps_handler import APPSTaskHandler
+from .arc.arc_handler import ARCChallengeTaskHandler
+from .common import TaskHandler
+from .gpqa_diamond.gpqa_diamond_handler import GPQADiamondTaskHandler
+from .gsm8k.gsm8k_handler import GSM8KTaskHandler
+from .livecodebench.livecodebench_handler import LiveCodeBenchTaskHandler
+from .math.math_handler import MATH500TaskHandler, MathTaskHandler
+from .mmlu.mmlu_handler import MMLUProTaskHandler, MMLUTaskHandler
+from .numina.numina_handler import NUMINATaskHandler
+from .taco.taco_handler import TACOTaskHandler
+
+TASK_HANDLERS = {
+    "NUMINA": NUMINATaskHandler,
+    "APPS": APPSTaskHandler,
+    "TACO": TACOTaskHandler,
+    "MATH500": MATH500TaskHandler,
+    "AIME": AIMETaskHandler,
+    "GPQADiamond": GPQADiamondTaskHandler,
+    "MMLU": MMLUTaskHandler,
+    "MMLUPro": MMLUProTaskHandler,
+    "LiveCodeBench": LiveCodeBenchTaskHandler,
+    "GSM8K": GSM8KTaskHandler,
+    "ARC-C": ARCChallengeTaskHandler,
+    "AMC23": AMC23TaskHandler,
+}
+
+
+__all__ = [
+    AIMETaskHandler,
+    APPSTaskHandler,
+    TACOTaskHandler,
+    MATH500TaskHandler,
+    AMC23TaskHandler,
+    NUMINATaskHandler,
+    GPQADiamondTaskHandler,
+    MMLUTaskHandler,
+    MMLUProTaskHandler,
+    LiveCodeBenchTaskHandler,
+    GSM8KTaskHandler,
+    ARCChallengeTaskHandler,
+    TaskHandler,
+    MathTaskHandler,
+    TASK_HANDLERS,
+]
